// frontend/src/types/RestaurantDetails.ts
export type RestaurantDetails = {
  id: number;
  name: string;
  description: string;
  address: string;
  email: string;
  operating_hours: Record<string, string>;
  booking_duration: number;
  tables_capacity: number;
  cuisine: string;
  logo_url?: string;
  menu_url?: string;
  phone: string;
  inst_url: string;
  rating: string;
  is_pet_friendly: boolean;
  created_at: string;
  updated_at: string;
  deleted_at: string | null;
};